# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OepenTrench API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/online-learning/disagreements": {
      "get": {
        "tags": [
          "online-learning"
        ],
        "summary": "List Disagreements",
        "operationId": "list_disagreements_online_learning_disagreements_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlineLearningDisagreementsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/online-learning/trainings": {
      "get": {
        "tags": [
          "online-learning"
        ],
        "summary": "List Trainings",
        "operationId": "list_trainings_online_learning_trainings_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlineLearningTrainingRunsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "online-learning"
        ],
        "summary": "Start Training",
        "operationId": "start_training_online_learning_trainings_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlineLearningTrainingRunRead"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GpsCoordinates": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Point",
            "title": "Type",
            "default": "Point"
          },
          "coordinates": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "GpsCoordinates"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OnlineLearningDisagreementsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OnlineLearningMismatchItemRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "stats": {
            "$ref": "#/components/schemas/OnlineLearningStatsRead"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "stats"
        ],
        "title": "OnlineLearningDisagreementsPage"
      },
      "OnlineLearningMismatchItemRead": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "original_label": {
            "type": "string",
            "title": "Original Label"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reviewed At"
          },
          "analysis": {
            "$ref": "#/components/schemas/PhotoAnalysisRead"
          },
          "mismatch_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mismatch Fields"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "project_id",
          "project_name",
          "original_label",
          "created_at",
          "reviewed_at",
          "analysis",
          "mismatch_fields"
        ],
        "title": "OnlineLearningMismatchItemRead"
      },
      "OnlineLearningStatsRead": {
        "properties": {
          "total_reviewed": {
            "type": "integer",
            "title": "Total Reviewed"
          },
          "total_mismatch": {
            "type": "integer",
            "title": "Total Mismatch"
          },
          "mismatch_rate": {
            "type": "number",
            "title": "Mismatch Rate"
          },
          "projects_with_mismatch": {
            "type": "integer",
            "title": "Projects With Mismatch"
          }
        },
        "type": "object",
        "required": [
          "total_reviewed",
          "total_mismatch",
          "mismatch_rate",
          "projects_with_mismatch"
        ],
        "title": "OnlineLearningStatsRead"
      },
      "OnlineLearningTrainingRunRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/OnlineLearningTrainingStatus"
          },
          "photo_count": {
            "type": "integer",
            "title": "Photo Count"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "photo_count",
          "started_at",
          "finished_at",
          "duration_sec",
          "error_message"
        ],
        "title": "OnlineLearningTrainingRunRead"
      },
      "OnlineLearningTrainingRunsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OnlineLearningTrainingRunRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "OnlineLearningTrainingRunsPage"
      },
      "OnlineLearningTrainingStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed"
        ],
        "title": "OnlineLearningTrainingStatus"
      },
      "PhotoAnalysisRead": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "is_in_domain": {
            "type": "boolean",
            "title": "Is In Domain"
          },
          "has_white_paper": {
            "type": "boolean",
            "title": "Has White Paper"
          },
          "has_ruler": {
            "type": "boolean",
            "title": "Has Ruler"
          },
          "estimated_depth": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Depth"
          },
          "has_duct": {
            "type": "boolean",
            "title": "Has Duct"
          },
          "estimate_number_of_ducts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimate Number Of Ducts"
          },
          "has_gdpr_problems": {
            "type": "boolean",
            "title": "Has Gdpr Problems"
          },
          "is_duplicated": {
            "type": "boolean",
            "title": "Is Duplicated"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PhotoDocumentationCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "gps_matches_route": {
            "type": "boolean",
            "title": "Gps Matches Route"
          },
          "date_valid": {
            "type": "boolean",
            "title": "Date Valid"
          },
          "is_false_call": {
            "type": "boolean",
            "title": "Is False Call"
          },
          "reviewer_override_category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PhotoDocumentationCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewer_has_duct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Has Duct"
          },
          "reviewer_has_ruler": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Has Ruler"
          },
          "reviewer_is_in_domain": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Is In Domain"
          },
          "reviewer_has_gdpr_problems": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Has Gdpr Problems"
          },
          "reviewer_gps_matches_route": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Gps Matches Route"
          },
          "reviewer_is_duplicated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Is Duplicated"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "gps_coordinates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GpsCoordinates"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "effective_has_duct": {
            "type": "boolean",
            "title": "Effective Has Duct"
          },
          "effective_has_ruler": {
            "type": "boolean",
            "title": "Effective Has Ruler"
          },
          "effective_is_in_domain": {
            "type": "boolean",
            "title": "Effective Is In Domain"
          },
          "effective_has_gdpr_problems": {
            "type": "boolean",
            "title": "Effective Has Gdpr Problems"
          },
          "effective_gps_matches_route": {
            "type": "boolean",
            "title": "Effective Gps Matches Route"
          },
          "effective_is_duplicated": {
            "type": "boolean",
            "title": "Effective Is Duplicated"
          },
          "effective_category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PhotoDocumentationCategory"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "is_in_domain",
          "has_white_paper",
          "has_ruler",
          "estimated_depth",
          "has_duct",
          "estimate_number_of_ducts",
          "has_gdpr_problems",
          "is_duplicated",
          "category",
          "gps_matches_route",
          "date_valid",
          "is_false_call",
          "reviewer_override_category",
          "reviewer_has_duct",
          "reviewer_has_ruler",
          "reviewer_is_in_domain",
          "reviewer_has_gdpr_problems",
          "reviewer_gps_matches_route",
          "reviewer_is_duplicated",
          "reviewed_at",
          "gps_coordinates",
          "created_at",
          "updated_at",
          "effective_has_duct",
          "effective_has_ruler",
          "effective_is_in_domain",
          "effective_has_gdpr_problems",
          "effective_gps_matches_route",
          "effective_is_duplicated",
          "effective_category"
        ],
        "title": "PhotoAnalysisRead"
      },
...
      "PhotoDocumentationCategory": {
        "type": "string",
        "enum": [
          "green",
          "yellow",
          "red"
        ],
        "title": "PhotoDocumentationCategory",
        "description": "Per-photo documentation quality (map node/segment rollup), not workflow status.\n\nRules and review workflow: docs/photo-documentation-category.md"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}